@@ -1,21 +1,15 @@
-class Leaderboard < ActiveRecord::Base
-  has_many :activities
-  before_destroy :detach_dependent_activities
-
+class Leaderboard < ApplicationRecord
+  has_many :activities, dependent: :nullify
 
   class << self
-
     def calc_new_leaderboards(new_activities)
-      new_activities.each do |activity|
-        update_leaderboard_with activity
-      end
+      new_activities.each { |activity| update_leaderboard_with activity }
     end
 
     def get_leaderboard_for_day(date=Time.zone.now.beginning_of_day.utc)
       Leaderboard
           .where(created_at: date.midnight..date.end_of_day)
-          .sort_by { |l| l.calories }
-          .reverse
+          .order(calories: :desc)
     end
 
     ################################################
@@ -23,39 +17,27 @@ class Leaderboard < ActiveRecord::Base
       def update_leaderboard_with(activity)
         l = Leaderboard.where(athlete_id: activity.athlete_id,
                               created_at: activity.start_date_local.beginning_of_day..activity.start_date_local.end_of_day).take
-        if l.nil?
-          l = create_new_entry(activity)
-        elsif l.activities.include? activity
-          continue
-        else
-          l = update_entry(activity, l)
-        end
+        return if l && l.activities.include?(activity)
+        l = l ? update_entry(activity, l) : create_new_entry(activity)
         l.activities << activity
-        l.save
       end
 
       def update_entry(activity, l)
-        l.calories += activity.calories
-        l.beers    = Activity.calc_beers(l.calories)
-        l
+        new_calories = l.calories + activity.calories
+        l.update_attributes(
+          calories: new_calories,
+          beers: Activity.calc_beers(new_calories)
+        )
+        l.reload
       end
 
       def create_new_entry(activity)
         athlete = Athlete.find(activity.athlete_id)
-        Leaderboard.new(athlete_name: athlete.name, athlete_id: athlete.id, img_url: athlete.img_url,
-                                  beers: activity.beers, calories: activity.calories, activity_id: activity.id,
-                                  created_at: activity.start_date_local)
+        Leaderboard.create!(
+          athlete_name: athlete.name, athlete_id: athlete.id,
+          img_url: athlete.img_url, beers: activity.beers,
+          calories: activity.calories, created_at: activity.start_date_local
+        )
       end
   end
-
-  #################################################
-  private
-    def detach_dependent_activities
-      if activities
-        activities.each do |a|
-          a.leaderboard_id = nil
-          a.save
-        end
-      end
-    end
 end
