@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,64 +10,58 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160706053514) do
+ActiveRecord::Schema.define(version: 20171027050112) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "activities", force: :cascade do |t|
-    t.string   "name"
-    t.float    "distance"
-    t.string   "activity_type"
-    t.integer  "moving_time"
-    t.float    "total_elevation_gain"
-    t.integer  "calories"
-    t.float    "start_lat"
-    t.float    "start_long"
-    t.float    "end_lat"
-    t.float    "end_long"
-    t.integer  "kudos_count"
-    t.datetime "created_at",           null: false
-    t.datetime "updated_at",           null: false
-    t.integer  "athlete_id"
-    t.integer  "beers"
-    t.string   "timezone"
-    t.datetime "start_date_local"
-    t.integer  "leaderboard_id"
-  end
-
-  add_index "activities", ["athlete_id", "created_at"], name: "index_activities_on_athlete_id_and_created_at", using: :btree
-  add_index "activities", ["athlete_id"], name: "index_activities_on_athlete_id", using: :btree
-  add_index "activities", ["leaderboard_id", "created_at"], name: "index_activities_on_leaderboard_id_and_created_at", using: :btree
-  add_index "activities", ["leaderboard_id"], name: "index_activities_on_leaderboard_id", using: :btree
-
-  create_table "athletes", force: :cascade do |t|
-    t.string   "name"
-    t.decimal  "calories"
-    t.decimal  "beers"
-    t.string   "img_url"
+  create_table "activities", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.float "distance"
+    t.string "activity_type"
+    t.integer "moving_time"
+    t.float "total_elevation_gain"
+    t.integer "calories"
+    t.float "start_lat"
+    t.float "start_long"
+    t.float "end_lat"
+    t.float "end_long"
+    t.integer "kudos_count"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.string   "username"
+    t.integer "athlete_id"
+    t.integer "beers"
+    t.string "timezone"
+    t.datetime "start_date_local"
+    t.bigint "leaderboard_id"
+    t.index ["athlete_id", "created_at"], name: "index_activities_on_athlete_id_and_created_at"
+    t.index ["athlete_id"], name: "index_activities_on_athlete_id"
+    t.index ["leaderboard_id"], name: "index_activities_on_leaderboard_id"
   end
 
-  create_table "leaderboards", force: :cascade do |t|
-    t.datetime "created_at",   null: false
-    t.datetime "updated_at",   null: false
-    t.string   "athlete_name"
-    t.string   "img_url"
-    t.decimal  "beers"
-    t.integer  "calories"
-    t.integer  "activity_id"
-    t.integer  "athlete_id"
+  create_table "athletes", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.decimal "calories"
+    t.decimal "beers"
+    t.string "img_url"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "username"
   end
 
-  add_index "leaderboards", ["activity_id"], name: "index_leaderboards_on_activity_id", using: :btree
-  add_index "leaderboards", ["athlete_id"], name: "index_leaderboards_on_athlete_id", using: :btree
-  add_index "leaderboards", ["created_at"], name: "index_leaderboards_on_created_at", using: :btree
+  create_table "leaderboards", id: :serial, force: :cascade do |t|
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "athlete_name"
+    t.string "img_url"
+    t.decimal "beers"
+    t.integer "calories"
+    t.integer "athlete_id"
+    t.index ["athlete_id"], name: "index_leaderboards_on_athlete_id"
+    t.index ["created_at"], name: "index_leaderboards_on_created_at"
+  end
 
   add_foreign_key "activities", "athletes"
   add_foreign_key "activities", "leaderboards"
-  add_foreign_key "leaderboards", "activities"
   add_foreign_key "leaderboards", "athletes"
 end
